fix: Fixed dashboard API call.
Fixed role icon appbar colour
Added cascade delete to tables for delete team.
